backend/requirements-fixed.txt
# Core
fastapi==0.95.0
uvicorn[standard]==0.21.1
python-socketio[asyncio]==5.8.0
python-multipart==0.0.6
pydantic==1.10.7

# Database
sqlalchemy==2.0.9
alembic==1.11.1
psycopg2-binary==2.9.6

# AI/ML
tensorflow-cpu==2.10.0
numpy==1.23.5
Pillow==9.5.0

# Async
anyio==3.6.2
httpx==0.24.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task Queue
celery[redis]==5.2.7
redis==4.5.4

# Environment
python-dotenv==0.21.1

# Google AI
google-generativeai==0.3.2
